Fix memory leak in image component
