+++ apps/web/hooks/useIsMounted.ts
@@ -1,0 +1,13 @@
+import { useEffect, useRef } from 'react'
+
+const useIsMounted = () => {
+  const isMounted = useRef(true)
+  useEffect(() => {
+    return () => {
+      isMounted.current = false
+    }
+  }, [])
+  return isMounted
+}
+
+export default useIsMounted
