+++ libs/island-ui/contentful/src/lib/Image/Image.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useState, useEffect } from 'react'
+import React, { FC, useState, useEffect, useRef } from 'react'
 import cn from 'classnames'
 import * as styles from './Image.treat'
 
@@ -12,12 +12,27 @@ export interface ImageProps {
   height: number
 }
 
+const useIsMounted = () => {
+  const isMounted = useRef(true)
+  useEffect(() => {
+    return () => {
+      isMounted.current = false
+    }
+  }, [])
+  return isMounted
+}
+
 const useImageLoader = (url: string): boolean => {
+  const isMounted = useIsMounted()
   const [loaded, setLoaded] = useState(false)
 
   useEffect(() => {
     const img = new window.Image(100)
-    img.onload = img.onerror = () => setLoaded(true)
+    img.onload = img.onerror = () => {
+      if (isMounted.current) {
+        setLoaded(true)
+      }
+    }
     img.src = url
   }, [url])
